chore: move internal function to utils
